AI-Personal-Coach/requirements.txt
openai
streamlit
transformers
nltk
fastapi
sqlite3
speechrecognition
gtts
python-dotenv
